requirements.txt
accelerate==0.22.0
datasets==2.14.4
diffusers==0.21.0.dev0
huggingface-hub==0.16.4
matplotlib==3.7.2
multidict==6.0.4
multiprocess==0.70.15
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchvision==0.11.2+cu111
